backend/requirements.txt
Flask>=2.0
pdfplumber
python-docx
openai
python-dotenv
